+++ packages/input-tag/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "description": "Illa Design InputTag.",
   "keywords": [
-    "input",
+    "input-tag",
     "react",
     "component"
   ],
