+++ packages/input-tag/src/utils.ts
@@ -1,4 +1,3 @@
-import { ReactNode } from "react"
 import { isArray, isObject } from "@illa-design/system"
 import { ObjectValueType } from "./interface"
